% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introns.R
\name{group_introns}
\alias{group_introns}
\title{Group intron by shared sites}
\usage{
group_introns(int_counts, cells)
}
\arguments{
\item{int_counts}{output of \code{\link{load_introns}}}

\item{cells}{dataframe with cell metadata, for example from cell_meta.rds created by nf-scsajr}
}
\value{
RangedSummarizedExperiment
}
\description{
The feature is splicing site*intron. Exlusion read count is sum of counts for all introns that share same site but not identical to the given.
}
\examples{
ints = load_introns('path/to/rds',c('sid1','sid2'))
cells = readRDS('path/to/rds/cell_meta.rds')
group_introns(ints,)
}
