% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introns.R
\name{load_introns}
\alias{load_introns}
\title{Load all intron counts from rds files outputed by nf-scsajr}
\usage{
load_introns(path2rds, sids)
}
\arguments{
\item{path2rds}{path to rds folder (nf-scsajr output)}

\item{sids}{vector of sample ifs}
}
\value{
combined dgCMatrix matrix
}
\description{
Load all intron counts from rds files outputed by nf-scsajr
}
\examples{
load_introns('path/to/rds',c('sid1','sid2'))
}
